test(ast): move traversal tests to session
